+++ crates/enoki2d/src/particles_2d/sprite.rs
@@ -1,10 +1,5 @@
 use super::{Particle2dMaterial, PARTICLE_SPRITE_FRAG};
-use bevy::{
-    asset::{io::Reader, AssetLoadError, AssetLoader, AsyncReadExt},
-    prelude::*,
-    render::render_resource::AsBindGroup,
-};
-use serde::Deserialize;
+use bevy::{prelude::*, render::render_resource::AsBindGroup};
 
 /// Sprite Material lets you add textures and animations
 /// to particles.
@@ -33,43 +28,17 @@ impl SpriteParticle2dMaterial {
             frame_data: UVec4::new(max_hframes, max_vframes, 0, 0),
         }
     }
+
+    pub fn from_texture(texture: Handle<Image>) -> Self {
+        Self {
+            texture: Some(texture),
+            frame_data: UVec4::new(1, 1, 0, 0),
+        }
+    }
 }
 
 impl Particle2dMaterial for SpriteParticle2dMaterial {
     fn fragment_shader() -> bevy::render::render_resource::ShaderRef {
         PARTICLE_SPRITE_FRAG.into()
-    }
-}
-
-#[derive(Deserialize)]
-pub struct SpriteParticle2dRon {
-    pub image: String,
-    pub h_frames: u32,
-    pub v_frames: u32,
-}
-
-#[derive(Default)]
-pub struct ColorParticle2dAssetLoader;
-impl AssetLoader for ColorParticle2dAssetLoader {
-    type Asset = SpriteParticle2dMaterial;
-    type Settings = ();
-    type Error = AssetLoadError;
-
-    async fn load(
-        &self,
-        reader: &mut dyn Reader,
-        _settings: &Self::Settings,
-        load_context: &mut bevy::asset::LoadContext<'_>,
-    ) -> Result<Self::Asset, Self::Error> {
-        let mut bytes = Vec::new();
-        reader.read_to_end(&mut bytes).await.unwrap();
-        let ron_asset = ron::de::from_bytes::<SpriteParticle2dRon>(bytes.as_slice())
-            .map_err(|_| AssetLoadError::AssetMetaReadError)?;
-
-        Ok(SpriteParticle2dMaterial::new(
-            load_context.load(ron_asset.image),
-            ron_asset.h_frames,
-            ron_asset.v_frames,
-        ))
     }
 }